fix(non-Windows): _FileMetadata Int64 cast + _NIOFileSystemFoundationCompat import
Caught by swiftci's CI on 2026-05-18 (macOS-15, Linux Swift 6.1):

1. Sources/Vapor/Utilities/FileMetadata.swift:37: cannot convert

   'Int' to 'Int64' — Timespec.seconds is Int on non-Windows but the

   _FileMetadata.lastModifiedSeconds field is Int64. Cast explicitly.

2. Sources/Vapor/Utilities/FileMetadata.swift:36: property 'date' is

   not available due to missing import of defining module

   '_NIOFileSystemFoundationCompat'. Swift 6.1's stricter

   MemberImportVisibility diagnostic. Add the explicit import on

   non-Windows (the product is already a non-Windows dependency).

+++ Sources/Vapor/Utilities/FileMetadata.swift
@@ -9,6 +9,12 @@
 import Foundation
 #if !os(Windows)
 import _NIOFileSystem
+// `_NIOFileSystemFoundationCompat` provides `FileInfo.Timespec.date` via
+// Foundation bridging. Swift 6.1's stricter `MemberImportVisibility`
+// diagnostic requires the defining module to be imported explicitly even
+// though `_NIOFileSystem` re-exports it. (Swift 6.3 tolerates the
+// implicit form.)
+import _NIOFileSystemFoundationCompat
 #endif
 
 /// Internal struct giving FileIO and FileMiddleware a single API for "did the path exist,
@@ -34,7 +40,11 @@ internal struct _FileMetadata: Sendable {
         return _FileMetadata(
             size: info.size,
             lastModifiedDate: info.lastDataModificationTime.date,
-            lastModifiedSeconds: info.lastDataModificationTime.seconds,
+            // `_NIOFileSystem.FileInfo.Timespec.seconds` is `Int` on
+            // macOS/Linux (where `time_t` is platform-`Int`) but the field
+            // is `Int64` on Windows (where `time_t` is `__time64_t`).
+            // Explicit `Int64(...)` keeps both platforms happy.
+            lastModifiedSeconds: Int64(info.lastDataModificationTime.seconds),
             isDirectory: info.type == .directory
         )
         #endif
